refactor: rename mem agent task to hotpotqa
