+++ examples/quickstart/training/task_config_hotpotqa.yaml
@@ -1,4 +1,4 @@
-- name: mem_agent
+- name: hotpotqa
   sandbox:
     provider: local
   agent:
