--- uni_agent/tasks/mem_agent/__init__.py
@@ -1,8 +1,0 @@
-"""MemAgent task and reward."""
-
-from __future__ import annotations
-
-from .reward import compute_score
-from .task import MemAgentTask, MemAgentTaskConfig
-
-__all__ = ["MemAgentTask", "MemAgentTaskConfig", "compute_score"]
